lesson-5 done

+++ src/types/product/product.resolvers.js
@@ -11,6 +11,9 @@ const productsTypeMatcher = {
 
 /** product */
 const product = (_, args, ctx) => {
+  if (!ctx.user) {
+    throw new AuthenticationError()
+  }
   return Product.findById(args.id)
     .lean()
     .exec()
@@ -18,17 +21,26 @@ const product = (_, args, ctx) => {
 
 const newProduct = (_, args, ctx) => {
   // use this fake ID for createdBy for now until we talk auth
+  if (!ctx.user || ctx.user.role !== roles.admin) {
+    throw new AuthenticationError()
+  }
   const createdBy = mongoose.Types.ObjectId()
-  return Product.create({ ...args.input, createdBy })
+  return Product.create({ ...args.input, createdBy: ctx.user._id })
 }
 
 const products = (_, args, ctx) => {
+  if (!ctx.user) {
+    throw new AuthenticationError()
+  }
   return Product.find({})
     .lean()
     .exec()
 }
 
 const updateProduct = (_, args, ctx) => {
+  if (!ctx.user || ctx.user.role !== roles.admin) {
+    throw new AuthenticationError()
+  }
   const update = args.input
   return Product.findByIdAndUpdate(args.id, update, { new: true })
     .lean()
@@ -36,6 +48,9 @@ const updateProduct = (_, args, ctx) => {
 }
 
 const removeProduct = (_, args, ctx) => {
+  if (!ctx.user || ctx.user.role !== roles.admin) {
+    throw new AuthenticationError()
+  }
   return Product.findByIdAndRemove(args.id)
     .lean()
     .exec()
